flowfile_core/requirements.txt
matplotlib~=3.9.2
fastapi ~= 0.115.2
uvicorn ~= 0.32.0
pydantic ~= 2.9.2
databases ~= 0.9.0
passlib ~= 1.7.4
pymysql ~= 1.1.1
pyarrow ~= 17.0.0
python-multipart ~= 0.0.12
polars ~= 1.9.0
methodtools ~= 0.4.7
faker ~= 23.1.0
XlsxWriter ~= 3.2.0
loky~=3.4.1
fastexcel~=0.12.0
openpyxl~=3.1.2
polars-distance~=0.4.3
polars-expr-transformer~=0.3.8
gspread~=6.1.3
google-api-python-client~=2.149.0
polars-sim~=0.2.3
polars-grouper~=0.2.0
polars-ds~=0.6.0
airbyte~=0.18.1
airbyte-cdk~=5.14.0